Add Ziggy to Inertia shared data
Ziggy has been imported and added to the shared props in HandleInertiaRequests middleware. This provides ability to use Ziggy's route helper in Inertia/Vue components. The location prop has also been included in the shared data to provide current URL to the components. The method comments for version() and share() have been cleaned.